+++ gpaw/test/big/g2_1/pbe_gpaw_nrel.agts.py
@@ -30,15 +30,16 @@ def agts(queue):
                        creates=[tag + '_ea_vs.csv'])
         plot = queue.add(tag + '_plot.py',
                          ncpus=1, walltime=5, deps=[analyse],
-                         creates=[tag + '.png'])
+                         creates=[tag + '_ea_vs.png'])
     # optimization
-    opt_run = [queue.add(tag + '_opt_run.py %d' % i,
-                         queueopts='-l nodes=1:ppn=2:opteron:ethernet',
-                         ncpus=1,
-                         walltime=40*60,
-                         deps=[run_generate])
-               for i in range(20)]
-    if 1:  # run after releasing new setups
+    if 0:  # run after releasing new setups
+        opt_run = [queue.add(tag + '_opt_run.py %d' % i,
+                             queueopts='-l nodes=1:ppn=2:opteron:ethernet',
+                             ncpus=1,
+                             walltime=40*60,
+                             deps=[run_generate])
+                   for i in range(20)]
+    if 0:  # run after releasing new setups
         opt_analyse = queue.add(tag + '_opt_analyse.py',
                                 ncpus=1, walltime=5, deps=opt_run,
                                 creates=[tag + '_opt_ea.csv',