+++ gpaw/test/big/scf/dcdft_pbe_pw.agts.py
@@ -7,12 +7,15 @@ runs = [
     #'d105', 'd107', 'd203',  # for MixerDif nmaxold does not matter
     'inititer10m', # cg steps and switching to rmm-diis does not help
     'cgm103', 'cgm203',  # cg helps
-    'cgs103',  # cg + MixerSum
+    #'cgs103',  # cg + MixerSum
     'cgd103', 'cgd203', 'cgd253', # cg + MixerDif: the winner
     'dzpm103',  # lcao guess does not help
-    'cgdzpm103',  # lcao + cg
-    'cgdzps103',  # lcao + cg + MixerSum
-    'cgdzpd103',  # lcao + cg + MixerDif
+    #'cgdzpm103',  # lcao + cg
+    #'cgdzps103',  # lcao + cg + MixerSum
+    #'cgdzpd103',  # lcao + cg + MixerDif
+    'fm103',  # FFT
+    'davd203',  # PW mode winner
+    'davfd203',  # FFT
     ]
 runsstr = ','.join(runs)
 runsstr += ',jacapo'
@@ -27,5 +30,5 @@ def agts(queue):
                        walltime=5*60)
     analyse = queue.add('analyse.py bulk scf_dcdft_pbe_pw ' + runsstr,
                         ncpus=1, walltime=10, deps=run + [jacapo],
-                        creates=['scf_dcdft_pbe_pbe_pw.csv',
-                                 'scf_dcdft_pbe_pbe_pw_calculator_steps.png'])
+                        creates=['scf_dcdft_pbe_pw_energy.csv',
+                                 'scf_dcdft_pbe_pw_calculator_steps.png'])